Improve the names of DoG op implementations
